Add matrix utils

+++ src/message.rs
@@ -16,10 +16,10 @@ pub static mut HISTORY_DIRECTION_Z: Vec<f64> = Vec::<f64>::new();
 
 pub struct Message {
     pub position_updated: bool,
-    pub position: Vector3<f32>,
-    pub acceleration: Vector3<f32>,
-    pub direction: Vector3<f32>,
-    pub initial_speed: f32,
+    pub position: Vector3<f64>,
+    pub acceleration: Vector3<f64>,
+    pub direction: Vector3<f64>,
+    pub initial_speed: f64,
 }
 
 impl Message {
@@ -52,31 +52,31 @@ impl Message {
 
             if buf.contains("POSITION") {
                 msg.position = Vector3::new(
-                    splitted[1].parse::<f32>().unwrap(), 
-                    splitted[2].parse::<f32>().unwrap(), 
-                    splitted[3].parse::<f32>().unwrap());
+                    splitted[1].parse::<f64>().unwrap(), 
+                    splitted[2].parse::<f64>().unwrap(), 
+                    splitted[3].parse::<f64>().unwrap());
                 HISTORY_POSITION_X.push(msg.position.x as f64);
                 HISTORY_POSITION_Y.push(msg.position.y as f64);
                 HISTORY_POSITION_Z.push(msg.position.z as f64);
                 msg.position_updated = true;
             } else if buf.contains("ACCELERATION") {
                 msg.acceleration = Vector3::new(
-                    splitted[1].parse::<f32>().unwrap(), 
-                    splitted[2].parse::<f32>().unwrap(),
-                    splitted[3].parse::<f32>().unwrap());
+                    splitted[1].parse::<f64>().unwrap(), 
+                    splitted[2].parse::<f64>().unwrap(),
+                    splitted[3].parse::<f64>().unwrap());
                 HISTORY_ACCELERATION_X.push(msg.acceleration.x as f64);
                 HISTORY_ACCELERATION_Y.push(msg.acceleration.y as f64);
                 HISTORY_ACCELERATION_Z.push(msg.acceleration.z as f64);
             } else if buf.contains("DIRECTION") {
                 msg.direction = Vector3::new(
-                    splitted[1].parse::<f32>().unwrap(), 
-                    splitted[2].parse::<f32>().unwrap(), 
-                    splitted[3].parse::<f32>().unwrap());
+                    splitted[1].parse::<f64>().unwrap(), 
+                    splitted[2].parse::<f64>().unwrap(), 
+                    splitted[3].parse::<f64>().unwrap());
                 HISTORY_DIRECTION_X.push(msg.direction.x as f64);
                 HISTORY_DIRECTION_Y.push(msg.direction.y as f64);
                 HISTORY_DIRECTION_Z.push(msg.direction.z as f64);
             } else if buf.contains("SPEED") {
-                msg.initial_speed = splitted[1].parse::<f32>().unwrap();
+                msg.initial_speed = splitted[1].parse::<f64>().unwrap();
             }
             buf = recv_message(socket);
         }
